+++ zapit/+zapit/+gui/+main/@controller/updateTestSiteDropdown.m
@@ -1,7 +1,7 @@
 function updateTestSiteDropdown(obj)
     % Updates drop-down menu showing stim locations so we can test-zap
     %
-    % 
+    %
     if isempty(obj.model.stimConfig)
         obj.TestSiteDropDown.Items = {};
         return
@@ -13,5 +13,5 @@ function updateTestSiteDropdown(obj)
 
     obj.TestSiteDropDown.Items = stimSite;
 
+    obj.TestSiteDropDown.ValueChangedFcn = @(~,~) obj.highlightSelectedStimCond;
 end % updateTestSiteDropDown
-