+++ CHANGELOG.md
@@ -29,6 +29,8 @@ NOTE: the version number of the software is taken from the last version number i
   * Minor: Produces a CLI warning message if sendSamples is called with a non-existent 
           condition number. 
   * BUGFIX: NI DAQ no longer requires DAQ IDs (names) to start with a "D". 
+  * The stimulus locations selected by the dropdown menu for interactive zapping are
+    now highlighted. 
 
 
 2023/10/31 v1.0.1